retrieve list of groups instead of jobs
