+++ src/lsst/cmservice/web_app/templates/campaign_card.html
@@ -16,8 +16,8 @@
             <div class="block text-xs pl-3">root: {{campaign.root}}</div>
             <div class="block text-xs pl-3">source: {{campaign.source}}</div>
 <!--        </p>-->
-        <div class="block text-xs pt-2 pl-3 text-green-700">{{campaign.jobs_completed}}</div>
         <div class="block text-xs pt-2 pl-3 text-green-700">{{campaign.scripts_completed}}</div>
+        <div class="block text-xs pt-2 pl-3 text-green-700">{{campaign.groups_completed}}</div>
     </div>
     <div>
         <span class="absolute right-3 bottom-3 text-teal-600" aria-hidden="true">
